Uses more XML features for clarity and ease of use with code.
Elements are now used. Whilst this will increase the size of the file, it will be more readable and easier to work with.
When values have an array of values (see proxyParameters and shaderArray), they will be stored within an element.

In addition,  this code supports shader filters. This has not been tested.